+++ projects/ng-container-terminal/src/lib/ct-yard/ct-yard.component.ts
@@ -1,20 +1,21 @@
-import { filter, bufferTime } from 'rxjs/operators';
 import {
   ChangeDetectionStrategy,
   Component,
   ElementRef,
+  EventEmitter,
   Input,
   OnChanges,
   OnInit,
-  SimpleChanges,
   Output,
-  EventEmitter
+  SimpleChanges,
 } from '@angular/core';
 import * as d3 from 'd3';
 import { Subject } from 'rxjs';
+import { bufferTime, filter } from 'rxjs/operators';
+
+import { RenderOptions } from '../../model/render-options';
 import { YardposInfo } from '../../model/yardpos-info';
 import { CtYardposParserService } from '../tool/ct-yardpos-parser.service';
-import { RenderOptions } from '../../model/render-options';
 
 @Component({
   selector: 'ct-yard',
@@ -37,7 +38,6 @@ export class CtYardComponent implements OnInit, OnChanges {
   maxTier = 5;
   maxBay = 0;
 
-  // TODO: 根据数据自动计算
   canvasWidth = 2000;
   canvasHeight = 100;
 
@@ -69,10 +69,7 @@ export class CtYardComponent implements OnInit, OnChanges {
   onNotifyDataChanged = new Subject<number>();
   onNotifyDataChanged$ = this.onNotifyDataChanged.asObservable();
 
-  constructor(
-    private el: ElementRef,
-    private yardposParser: CtYardposParserService
-  ) {}
+  constructor(private el: ElementRef, private yardposParser: CtYardposParserService) {}
 
   ngOnInit() {
     this.host = d3.select(this.el.nativeElement);
@@ -149,19 +146,26 @@ export class CtYardComponent implements OnInit, OnChanges {
    * pods: 所有卸货港数组
    */
   extractBasicInfo() {
-    this.maxRow = Math.max(
-      ...this.yardposInfoList.map(d => +this.yardposParser.getP(d.yardpos))
+    this.maxRow = Math.max(...this.yardposInfoList.map(d => +this.yardposParser.getP(d.yardpos)));
+    this.maxTier = Math.max(...this.yardposInfoList.map(d => +this.yardposParser.getC(d.yardpos)));
+    this.maxBay = Math.max(
+      ...this.yardposInfoList
+        .map(d => +this.yardposParser.getW(d.yardpos))
+        // .filter(bay => bay % 2 === 1)
+        .map(bay => (bay + (2 - (bay % 2))) / 2)
     );
-    this.maxTier = Math.max(
-      ...this.yardposInfoList.map(d => +this.yardposParser.getC(d.yardpos))
-    );
-    this.maxBay = d3
-      .set(
-        this.yardposInfoList
-          .map(d => +this.yardposParser.getW(d.yardpos))
-          .filter(bay => bay % 2 === 1)
-      )
-      .values().length;
+    console.log({
+      maxBay: this.maxBay,
+      maxRow: this.maxRow,
+      maxTier: this.maxTier
+    });
+    // this.maxBay = d3
+    //   .set(
+    //     this.yardposInfoList
+    //       .map(d => +this.yardposParser.getW(d.yardpos))
+    //       .filter(bay => bay % 2 === 1)
+    //   )
+    //   .values().length;
     const containers = this.yardposInfoList
       .filter(pos => pos.containers && pos.containers.length > 0)
       .map(pos => pos.containers)
@@ -193,39 +197,44 @@ export class CtYardComponent implements OnInit, OnChanges {
       }
     });
 
+    const evenBayPoses = this.yardposInfoList.filter(pos => +this.yardposParser.getW(pos.yardpos) % 2 === 0);
+
+    const oddBayPoses = this.yardposInfoList.filter(pos => +this.yardposParser.getW(pos.yardpos) % 2 === 1);
+
+    const invalidOddBay = [];
     bayInfo.forEach((info, idx) => {
+      if (idx % 2 === 0) {
+        if (info && info.containerCount > 0) {
+          const poses = evenBayPoses.filter(pos => +this.yardposParser.getW(pos.yardpos) === idx);
+          this.displayYardposInfoList = [...poses, ...this.displayYardposInfoList];
+          invalidOddBay.push(idx + 1);
+          invalidOddBay.push(idx - 1);
+        }
+      }
+
       // 如果是基数贝，则向前向后找其偶数倍是否存在占位信息，若不存在则需要画该基数贝
-      if (idx % 2 === 1) {
-        if (
-          (bayInfo[idx + 1] === undefined ||
-            (bayInfo[idx + 1].containerCount === 0 &&
-              bayInfo[idx + 1].planCount === 0)) &&
-          (bayInfo[idx - 1] === undefined ||
-            (bayInfo[idx - 1].containerCount === 0 &&
-              bayInfo[idx - 1].planCount === 0))
-        ) {
-          const poses = this.yardposInfoList.filter(
-            pos => +this.yardposParser.getW(pos.yardpos) === idx
-          );
-          this.displayYardposInfoList = [
-            ...poses,
-            ...this.displayYardposInfoList
-          ];
-        }
-      } else {
-        if (
-          info.containerCount > 0 ||
-          info.planCount > 0 ||
-          info.taskCount > 0
-        ) {
-          const poses = this.yardposInfoList.filter(
-            pos => +this.yardposParser.getW(pos.yardpos) === idx
-          );
-          this.displayYardposInfoList = [
-            ...poses,
-            ...this.displayYardposInfoList
-          ];
-        }
+      // if (idx % 2 === 1) {
+      //   if (
+      //     (bayInfo[idx + 1] === undefined ||
+      //       (bayInfo[idx + 1].containerCount === 0 && bayInfo[idx + 1].planCount === 0)) &&
+      //     (bayInfo[idx - 1] === undefined ||
+      //       (bayInfo[idx - 1].containerCount === 0 && bayInfo[idx - 1].planCount === 0))
+      //   ) {
+      //     const poses = this.yardposInfoList.filter(pos => +this.yardposParser.getW(pos.yardpos) === idx);
+      //     this.displayYardposInfoList = [...poses, ...this.displayYardposInfoList];
+      //   }
+      // } else {
+      //   if (info.containerCount > 0 || info.planCount > 0 || info.taskCount > 0) {
+      //     const poses = this.yardposInfoList.filter(pos => +this.yardposParser.getW(pos.yardpos) === idx);
+      //     this.displayYardposInfoList = [...poses, ...this.displayYardposInfoList];
+      //   }
+      // }
+    });
+
+    bayInfo.forEach((info, idx) => {
+      if (idx % 2 === 1 && invalidOddBay.indexOf(idx) < 0) {
+        const poses = oddBayPoses.filter(pos => +this.yardposParser.getW(pos.yardpos) === idx);
+        this.displayYardposInfoList = [...poses, ...this.displayYardposInfoList];
       }
     });
   }
@@ -234,6 +243,16 @@ export class CtYardComponent implements OnInit, OnChanges {
    * 视图渲染
    */
   private redraw() {
+    const baseWidth = +this.baseWidth;
+    const baseHeight = +this.baseHeight;
+    if (+this.rotation === 0) {
+      this.canvasWidth = (this.maxRow * this.maxBay + 1) * (baseWidth + 2 * this.interval);
+      this.canvasHeight = (this.maxTier + 1) * (baseHeight + 2 * this.interval);
+    } else {
+      this.canvasWidth = (this.maxTier * this.maxBay + 1) * (baseWidth + 2 * this.interval);
+      this.canvasHeight = (this.maxRow + 1) * (baseHeight + 2 * this.interval);
+    }
+    this.svg.attr('width', `${this.canvasWidth}px`).attr('height', `${this.canvasHeight}px`);
     // 绘制列标签
     this.rowLabelsGroup.selectAll('g.row-label').remove();
     let rowLabels = this.rowLabelsGroup.selectAll('g.row-label');
@@ -267,9 +286,7 @@ export class CtYardComponent implements OnInit, OnChanges {
 
     // 绘制基数贝标签
     this.oddBayLabelsGroup.selectAll('g.odd-bay-label').remove();
-    const oddBayLabels = this.oddBayLabelsGroup
-      .selectAll('g.odd-bay-label')
-      .data(d3.range(0, this.maxBay * 2, 2));
+    const oddBayLabels = this.oddBayLabelsGroup.selectAll('g.odd-bay-label').data(d3.range(0, this.maxBay * 2, 2));
     const oddBayLabel = oddBayLabels
       .enter()
       .append('g')
@@ -304,12 +321,15 @@ export class CtYardComponent implements OnInit, OnChanges {
 
     const yardPoses = this.yardGroup
       .selectAll('g.yardpos')
-      .data(this.displayYardposInfoList, data => JSON.stringify(data));
+      .data(this.displayYardposInfoList, (data: YardposInfo) => data.yardpos);
 
     // 更新
-    yardPoses.transition().attr('transform', (posInfo: YardposInfo) => {
-      return this._transformFunction(posInfo);
-    });
+    yardPoses
+      .transition()
+      .duration(500)
+      .attr('transform', (posInfo: YardposInfo) => {
+        return this._transformFunction(posInfo);
+      });
     yardPoses
       .selectAll('path.cell')
       .transition()
@@ -323,33 +343,25 @@ export class CtYardComponent implements OnInit, OnChanges {
           width = this.baseWidth * 2;
         }
 
-        const baseRect = `M0 0 L${width} 0 L${width} ${this.baseHeight} L0 ${
-          this.baseHeight
-        } Z`;
+        const baseRect = `M0 0 L${width} 0 L${width} ${this.baseHeight} L0 ${this.baseHeight} Z`;
         let finalRect = baseRect;
         if (data.isLocked) {
           // 有封场则画X表示
-          finalRect =
-            finalRect +
-            ` M0 0 L${width} ${this.baseHeight} M${width} 0 L0 ${
-              this.baseHeight
-            }`;
+          finalRect = finalRect + ` M0 0 L${width} ${this.baseHeight} M${width} 0 L0 ${this.baseHeight}`;
         }
         if (data.displayedContainer && data.displayedContainer.task) {
           // 显示的集装箱有任务则画圈表示
           finalRect =
             finalRect +
-            ` M0 ${this.baseHeight / 2} A${width / 2} ${width /
-              2} 0 0 1 ${width} ${this.baseHeight / 2}` + // 上半圈
-            ` M${width} ${this.baseHeight / 2} A${width / 2} ${width /
-              2} 0 0 1 ${0} ${this.baseHeight / 2}`; // 下半圈
+            ` M0 ${this.baseHeight / 2} A${width / 2} ${width / 2} 0 0 1 ${width} ${this.baseHeight / 2}` + // 上半圈
+            ` M${width} ${this.baseHeight / 2} A${width / 2} ${width / 2} 0 0 1 ${0} ${this.baseHeight / 2}`; // 下半圈
         }
         return finalRect;
       })
       .attr('fill', (data: YardposInfo) => {
         return this._fillFunction(data);
       })
-      .attr('stroke', 'rgb(90,68,70)')
+      .attr('stroke', (posInfo: YardposInfo) => this.draw('stroke', posInfo))
       .attr('stroke-width', '1px');
 
     // 高箱 需要加一条粗线
@@ -357,25 +369,22 @@ export class CtYardComponent implements OnInit, OnChanges {
       .selectAll('path.ctn-height')
       .transition()
       .attr('d', (data: YardposInfo) => {
-        if (
-          data.displayedContainer &&
-          data.displayedContainer.height + '' === '9.6'
-        ) {
+        if (data.displayedContainer && data.displayedContainer.height + '' === '9.6') {
           let factor = 1;
           if (data.displayedContainer.size !== '20') {
             factor = 2;
           }
-          return `M0 2 L${this.baseWidth * factor} 2`;
+          return `M0.5 2.5 L${this.baseWidth * factor} 2.5`;
         } else {
-          return `M0 0 L${this.baseWidth} 0`;
+          return `M0.5 0 L${this.baseWidth} 0`;
         }
       })
       .attr('stroke', 'black')
       .attr('stroke-width', (data: YardposInfo) => {
-        if (
-          data.displayedContainer &&
-          data.displayedContainer.height + '' === '9.6'
-        ) {
+        if (!data.displayedContainer) {
+          return 0;
+        }
+        if (data.displayedContainer.height + '' === '9.6') {
           return 4;
         } else {
           return 1;
@@ -436,7 +445,6 @@ export class CtYardComponent implements OnInit, OnChanges {
 
     pos
       .transition()
-
       .duration(300)
       .ease(d3.easeCubicOut)
       .attr('transform', posInfo => {
@@ -459,59 +467,48 @@ export class CtYardComponent implements OnInit, OnChanges {
           width = this.baseWidth * 2;
         }
 
-        const baseRect = `M0 0 L${width} 0 L${width} ${this.baseHeight} L0 ${
-          this.baseHeight
-        } Z`;
+        const baseRect = `M0 0 L${width} 0 L${width} ${this.baseHeight} L0 ${this.baseHeight} Z`;
         let finalRect = baseRect;
         if (data.isLocked) {
           // 有封场则画X表示
-          finalRect =
-            finalRect +
-            ` M0 0 L${width} ${this.baseHeight} M${width} 0 L0 ${
-              this.baseHeight
-            }`;
+          finalRect = finalRect + ` M0 0 L${width} ${this.baseHeight} M${width} 0 L0 ${this.baseHeight}`;
         }
         if (data.displayedContainer && data.displayedContainer.task) {
           // 显示的集装箱有任务则画圈表示
           finalRect =
             finalRect +
-            ` M0 ${this.baseHeight / 2} A${width / 2} ${width /
-              2} 0 0 1 ${width} ${this.baseHeight / 2}` + // 上半圈
-            ` M${width} ${this.baseHeight / 2} A${width / 2} ${width /
-              2} 0 0 1 ${0} ${this.baseHeight / 2}`; // 下半圈
+            ` M0 ${this.baseHeight / 2} A${width / 2} ${width / 2} 0 0 1 ${width} ${this.baseHeight / 2}` + // 上半圈
+            ` M${width} ${this.baseHeight / 2} A${width / 2} ${width / 2} 0 0 1 ${0} ${this.baseHeight / 2}`; // 下半圈
         }
         return finalRect;
       })
       .attr('fill', data => {
         return this._fillFunction(data);
       })
-      .attr('stroke', 'rgb(90,68,70)')
-      .attr('stroke-width', '1px');
+      .attr('stroke', posInfo => this.draw('stroke', posInfo))
+      .attr('stroke-width', posInfo => this.draw('strokeWidth', posInfo));
 
     // 高箱 需要加一条粗线
     pos
       .append('path')
       .attr('class', 'ctn-height')
       .attr('d', data => {
-        if (
-          data.displayedContainer &&
-          data.displayedContainer.height + '' === '9.6'
-        ) {
+        if (data.displayedContainer && data.displayedContainer.height + '' === '9.6') {
           let factor = 1;
           if (data.displayedContainer.size !== '20') {
             factor = 2;
           }
-          return `M0 2 L${this.baseWidth * factor} 2`;
+          return `M0.5 2.5 L${this.baseWidth * factor} 2.5`; // 0.5是stroke-width的一半，不然高箱的线会超出边界
         } else {
-          return `M0 0 L${this.baseWidth} 0`;
+          return `M0.5 0 L${this.baseWidth} 0`;
         }
       })
       .attr('stroke', 'black')
       .attr('stroke-width', data => {
-        if (
-          data.displayedContainer &&
-          data.displayedContainer.height + '' === '9.6'
-        ) {
+        if (!data.displayedContainer) {
+          return 0;
+        }
+        if (data.displayedContainer.height + '' === '9.6') {
           return 4;
         } else {
           return 1;
@@ -579,5 +576,20 @@ export class CtYardComponent implements OnInit, OnChanges {
       y = this.baseHeight * (this.maxTier - tier);
     }
     return `translate(${x}, ${y})`;
+  }
+
+  private draw(which: string, yardInfo: YardposInfo) {
+    if (this.renderOptions && this.renderOptions[which]) {
+      switch (typeof this.renderOptions[which]) {
+        case 'string':
+          return this.renderOptions[which];
+        case 'number':
+          return `${this.renderOptions[which]}`;
+        default:
+          return this.renderOptions[which](yardInfo);
+      }
+    } else {
+      return null;
+    }
   }
 }