+++ projects/ng-container-terminal/src/lib/ct-yard/ct-yard.module.ts
@@ -1,5 +1,6 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+
 import { CtYardComponent } from './ct-yard.component';
 
 @NgModule({