+++ projects/ng-container-terminal/src/lib/tool/ct-vescell-parser.service.ts
@@ -2,7 +2,9 @@ import { Inject, Injectable } from '@angular/core';
 
 import { VESCELL_PARSER_CONFIG, VescellParserConfig } from './model/vescell-parser-config';
 
-@Injectable()
+@Injectable({
+  providedIn: 'root'
+})
 export class CtVescellParserService {
   private defaultConfig: VescellParserConfig = {
     pattern: 'BBBBLLCC'
@@ -14,13 +16,11 @@ export class CtVescellParserService {
     C: { s: 6, e: 8 }
   };
 
-  private finalConfig: VescellParserConfig = this.defaultConfig;
+  private finalConfig: VescellParserConfig;
 
-  constructor(
-    @Inject(VESCELL_PARSER_CONFIG) private config: VescellParserConfig
-  ) {
+  constructor(@Inject(VESCELL_PARSER_CONFIG) private config: VescellParserConfig) {
     // 合并用户提供的解析器配置
-    this.finalConfig = Object.assign(this.finalConfig, config);
+    this.finalConfig = Object.assign({}, this.defaultConfig, config);
     // 重新计算索引
     this._updateIdxMap();
   }
@@ -52,6 +52,10 @@ export class CtVescellParserService {
 
   getC(vescell: string) {
     return vescell.slice(this.idxMap['C']['s'], this.idxMap['C']['e']);
+  }
+
+  getLC(vescell: string) {
+    return this.getL(vescell) + this.getC(vescell);
   }
 
   isDeck(vescell: string): boolean {