+++ projects/ng-container-terminal/src/model/yard-bay.ts
@@ -1,16 +1,4 @@
-import { Observable, Subject } from 'rxjs';
-
 import { Yardpos } from './yardpos';
-import { YardposInfo } from './yardpos-info';
-
-export interface YardBayOld {
-  name: string;
-  maxRow: number;
-  maxTier: number;
-  yardposInfoArray: YardposInfo[];
-  dataUpdated?: Observable<any>;
-  dataUpdatedSource?: Subject<any>;
-}
 
 export interface YardBay<T> {
   name: string;