+++ projects/ng-container-terminal/src/model/yard-info.ts
@@ -1,12 +1,12 @@
 export interface YardInfo<T> {
   block: string;
-  direction: 'H' | 'V';
+  direction?: 'H' | 'V';
   maxBay: number;
   maxRow: number;
   maxTier: number;
-  x: number;
-  y: number;
-  width: number;
-  height: number;
+  x?: number;
+  y?: number;
+  width?: number;
+  height?: number;
   data?: T;
 }