+++ projects/ng-container-terminal/src/model/yardpos-info.ts
@@ -1,6 +1,5 @@
-import { Plan } from './plan';
-import { Task } from './task';
 import { Container } from './container';
+import { Plan } from './plan';
 
 export interface YardposInfo {
   displayedContainer: Container;